requirements.txt
flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
prometheus_client==0.17.1
